+++ src/HDF5er/__init__.py
@@ -1,2 +1,3 @@
 from .MDA2HDF5 import *
 from .ase2HDF5 import *
+from .HDF5To import *