+++ test/test_HDFer.py
@@ -3,11 +3,10 @@ import h5py
 import pytest
 import MDAnalysis
 import numpy
+from MDAnalysis.lib.mdamath import triclinic_vectors
 
-# from MDAnalysis.tests.datafiles import
 
-
-def giveUniverse() -> MDAnalysis.Universe:
+def giveUniverse(angles: set = (90.0, 90.0, 90.0)) -> MDAnalysis.Universe:
     traj = numpy.array(
         [
             [[0.0, 0.0, 0.0], [1.0, 1.0, 1.0], [2.0, 2.0, 2.0], [3.0, 3.0, 3.0]],
@@ -25,7 +24,9 @@ def giveUniverse() -> MDAnalysis.Universe:
         traj,
         order="fac",
         # this tests the non orthogonality of the box
-        dimensions=numpy.array([[6.0, 6.0, 6.0, 90, 60, 90]] * traj.shape[0]),
+        dimensions=numpy.array(
+            [[6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]] * traj.shape[0]
+        ),
     )
     return u
 
@@ -50,6 +51,7 @@ def test_MDA2HDF5():
         # this checks also that the dataset has been created
         nat = len(group["Types"])
         assert len(group["Trajectory"]) == len(fourAtomsFiveFrames.trajectory)
+        assert len(group["Trajectory"]) == 5
         for i, f in enumerate(fourAtomsFiveFrames.trajectory):
             assert nat == len(fourAtomsFiveFrames.atoms)
             for atomID in range(nat):
@@ -92,20 +94,231 @@ def test_MDA2HDF5Sliced():
                         < 1e-8
                     )
 
-
-def test_ase():
     # this test has been writtend After the function has been written
     # This creates or overwite the test file:
-    fourAtomsFiveFrames = giveUniverse()
-    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+
+
+def test_MDA2HDF5Box():
+    fourAtomsFiveFrames = giveUniverse((90, 90, 90))
+    HDF5er.MDA2HDF5(
+        fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames_toOver", override=True
+    )
     with h5py.File("test.hdf5", "r") as hdf5test:
-        group = hdf5test["Trajectories/4Atoms5Frames"]
+        group = hdf5test["Trajectories/4Atoms5Frames_toOver"]
         aseTraj = HDF5er.HDF52AseAtomsChunckedwithSymbols(
             group, slice(5), slice(5), ["H", "H", "H", "H"]
         )
-        for i, d in enumerate([6.0, 0.0, 0.0]):
-            assert aseTraj[0].cell[0][i] - d < 1e-8
-        for i, d in enumerate([0.0, 6.0, 0.0]):
-            assert aseTraj[0].cell[1][i] - d < 1e-8
-        for i, d in enumerate(6.0 * numpy.array([0.5, 0.0, numpy.sqrt(3.0) / 2])):
-            assert aseTraj[0].cell[2][i] - d < 1e-8
+        for j, array in enumerate(triclinic_vectors([6.0, 6.0, 6.0, 90, 90, 90])):
+            for i, d in enumerate(array):
+                assert aseTraj[0].cell[j][i] - d < 1e-7
+    for angles in [
+        (90, 60, 90),
+        (60, 60, 60),
+        (50, 60, 90),
+        (90, 90, 45),
+        (80, 60, 90),
+    ]:
+        print(angles)
+        fourAtomsFiveFramesSkew = giveUniverse(angles)
+        HDF5er.MDA2HDF5(
+            fourAtomsFiveFramesSkew, "test.hdf5", "4Atoms5Frames_toOver", override=True
+        )
+        with h5py.File("test.hdf5", "r") as hdf5test:
+            group = hdf5test["Trajectories/4Atoms5Frames_toOver"]
+            aseTraj = HDF5er.HDF52AseAtomsChunckedwithSymbols(
+                group, slice(5), slice(5), ["H", "H", "H", "H"]
+            )
+            for j, array in enumerate(
+                triclinic_vectors(
+                    [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]],
+                    dtype=numpy.float64,
+                )
+            ):
+                for i, d in enumerate(array):
+                    assert aseTraj[0].cell[j][i] - d < 1e-7
+
+
+def test_copyMDA2HDF52xyz1DData():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    OneDData = rng.integers(
+        0, 7, size=(len(fourAtomsFiveFrames.trajectory), len(fourAtomsFiveFrames.atoms))
+    )
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(
+            group, OneDData=OneDData
+        )  # ,TwoDData=TwoDData)
+        # print(stringData)
+        lines = stringData.splitlines()
+        nat = int(lines[0])
+        assert int(lines[0]) == len(fourAtomsFiveFrames.atoms)
+        assert lines[2].split()[0] == fourAtomsFiveFrames.atoms.types[0]
+        for frame, traj in enumerate(fourAtomsFiveFrames.trajectory):
+            frameID = frame * (nat + 2)
+            assert int(lines[frameID]) == nat
+            t = lines[frameID + 1].split(" Lattice=")
+            Lattice = t[1].replace('"', "").split()
+            Properties = t[0].split(":")
+            # assert "TwoDData" in Properties
+            assert "OneDData" in Properties
+            for original, control in zip(latticeVector, Lattice):
+                assert (original - float(control)) < 1e-7
+            for atomID in range(len(fourAtomsFiveFrames.atoms)):
+                thisline = frameID + 2 + atomID
+                assert (
+                    lines[thisline].split()[0]
+                    == fourAtomsFiveFrames.atoms.types[atomID]
+                )
+                assert len(lines[thisline].split()) == 5
+                assert int((lines[thisline].split()[-1])) == OneDData[frame, atomID]
+                for i in range(3):
+                    assert (
+                        float(lines[thisline].split()[i + 1])
+                        == fourAtomsFiveFrames.atoms.positions[atomID][i]
+                    )
+
+
+def test_copyMDA2HDF52xyzMultiDData():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    dataDim = rng.integers(2, 15)
+    MultiDData = rng.integers(
+        0,
+        7,
+        size=(
+            len(fourAtomsFiveFrames.trajectory),
+            len(fourAtomsFiveFrames.atoms),
+            dataDim,
+        ),
+    )
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(group, MultiDData=MultiDData)
+        lines = stringData.splitlines()
+        nat = int(lines[0])
+        assert int(lines[0]) == len(fourAtomsFiveFrames.atoms)
+        assert lines[2].split()[0] == fourAtomsFiveFrames.atoms.types[0]
+        for frame, traj in enumerate(fourAtomsFiveFrames.trajectory):
+            frameID = frame * (nat + 2)
+            assert int(lines[frameID]) == nat
+            t = lines[frameID + 1].split(" Lattice=")
+            Lattice = t[1].replace('"', "").split()
+            Properties = t[0].split(":")
+            assert "MultiDData" in Properties
+            for original, control in zip(latticeVector, Lattice):
+                assert (original - float(control)) < 1e-7
+            for atomID in range(len(fourAtomsFiveFrames.atoms)):
+                thisline = frameID + 2 + atomID
+                assert (
+                    lines[thisline].split()[0]
+                    == fourAtomsFiveFrames.atoms.types[atomID]
+                )
+                assert len(lines[thisline].split()) == (4 + dataDim)
+                for i, d in enumerate(MultiDData[frame, atomID]):
+                    assert int((lines[thisline].split()[4 + i])) == d
+                for i in range(3):
+                    assert (
+                        float(lines[thisline].split()[i + 1])
+                        == fourAtomsFiveFrames.atoms.positions[atomID][i]
+                    )
+
+
+@pytest.mark.xfail(raises=ValueError)
+def test_copyMDA2HDF52xyz_error1D():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    OneDData = rng.integers(
+        0,
+        7,
+        size=(len(fourAtomsFiveFrames.trajectory), len(fourAtomsFiveFrames.atoms) + 1),
+    )
+
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(group, OneDData=OneDData)
+
+
+@pytest.mark.xfail(raises=ValueError)
+def test_copyMDA2HDF52xyz_error2D():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    TwoDData = rng.integers(
+        0,
+        7,
+        size=(
+            len(fourAtomsFiveFrames.trajectory),
+            len(fourAtomsFiveFrames.atoms) + 1,
+            2,
+        ),
+    )
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(group, TwoDData=TwoDData)
+
+
+@pytest.mark.xfail(raises=ValueError)
+def test_copyMDA2HDF52xyz_wrongD():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    WrongDData = rng.integers(
+        0,
+        7,
+        size=(
+            len(fourAtomsFiveFrames.trajectory),
+            len(fourAtomsFiveFrames.atoms),
+            2,
+            4,
+        ),
+    )
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(group, WrongDData=WrongDData)
+
+
+@pytest.mark.xfail(raises=ValueError)
+def test_copyMDA2HDF52xyz_wrongTrajlen():
+    angles = (75.0, 60.0, 90.0)
+    fourAtomsFiveFrames = giveUniverse(angles)
+    latticeVector = triclinic_vectors(
+        [6.0, 6.0, 6.0, angles[0], angles[1], angles[2]]
+    ).flatten()
+    rng = numpy.random.default_rng(12345)
+    WrongDData = rng.integers(
+        0,
+        7,
+        size=(
+            len(fourAtomsFiveFrames.trajectory) + 5,
+            len(fourAtomsFiveFrames.atoms),
+            2,
+        ),
+    )
+    HDF5er.MDA2HDF5(fourAtomsFiveFrames, "test.hdf5", "4Atoms5Frames", override=True)
+    with h5py.File("test.hdf5", "r") as hdf5test:
+        group = hdf5test["Trajectories/4Atoms5Frames"]
+        stringData = HDF5er.getXYZfromTrajGroup(group, WrongDData=WrongDData)